refactor pm quality fairness lookup routes
