+++ src/app/router_registry.py
@@ -64,6 +64,9 @@ from app.routers.dpm_command_center_pm_quality_ai import (
 from app.routers.dpm_command_center_pm_quality_fairness import (
     router as dpm_command_center_pm_quality_fairness_router,
 )
+from app.routers.dpm_command_center_pm_quality_fairness_lookup import (
+    router as dpm_command_center_pm_quality_fairness_lookup_router,
+)
 from app.routers.dpm_command_center_pm_quality_policies import (
     router as dpm_command_center_pm_quality_policies_router,
 )
@@ -246,6 +249,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_command_center_pm_quality_router)
     app.include_router(dpm_command_center_pm_quality_ai_router)
     app.include_router(dpm_command_center_pm_quality_fairness_router)
+    app.include_router(dpm_command_center_pm_quality_fairness_lookup_router)
     app.include_router(dpm_command_center_pm_quality_policies_router)
     app.include_router(dpm_command_center_pm_quality_review_actions_router)
     app.include_router(dpm_command_center_pm_quality_review_action_lookup_router)
